example/ticket-agent/requirements.txt
# Ticket Agent Dependencies
anthropic>=0.40.0
python-dotenv>=1.0.0
agentfense
